+++ projector_installer/apps.py
@@ -48,7 +48,12 @@ CHANNEL_SETTINGS_FILE = '.channel.settings.json'
 
 def get_installed_apps(pattern: Optional[str] = None) -> List[str]:
     """Returns sorted list of installed apps, matched given pattern."""
-    res = [file_name for file_name in listdir(get_apps_dir()) if
+    apps_dir = get_apps_dir()
+
+    if not isdir(apps_dir):
+        return []
+
+    res = [file_name for file_name in listdir(apps_dir) if
            pattern is None or file_name.lower().find(pattern.lower()) != -1]
     res.sort()
     return res